Added readMe, refactored logic throughout
