+++ ReadMe.md
@@ -1,0 +1,32 @@
+# ReadMe
+
+### Prerequisites -
+
+- MacOs
+- Node.js installed (if you do not already have this) you can run `node --version` to check this is installed.
+- A test file ready to go! 😄
+
+### Installation -
+
+- Clone this repository
+- In the root directory of the application, add your test file. ensure the file is called: `poker-hands.txt`
+- Navigate into the cloned repository
+
+in the terminal run the following commands -
+
+`npm install` This will install any dependencies incase your machine is missing native node packages.
+
+`npm install -g` This will install the poker checker globally, so you can run the CLI from any directory.
+
+**_You have now installed the CLI._**
+
+### Once you have finished the steps above -
+
+from anywhere within your machine directories run the following command -
+
+`run-poker-hands`
+
+- This will run the program
+- Once run, the program will write the test results to a .txt file called `game-result.txt` located in the `/gameResults` directory within the application folders. 🎉
+- the `game-result.txt` file will be empty originally.
+- The program will also log the result to the terminal.
