+++ gameResults/game-result.txt
@@ -1,2 +1,2 @@
-Player 1: 263 hands 
-Player 2: 237 hands
+Player 1: 12 hands 
+Player 2: 11 hands
