+++ package.json
@@ -1,5 +1,5 @@
 {
-  "name": "Argenti_Code_Challenge",
+  "name": "Poker Hands",
   "version": "1.0.0",
   "description": "",
   "main": "pokerHand.mjs",
@@ -14,7 +14,7 @@
     "url": "git+https://github.com/LukeMilsom/Poker-Hand-Sorter.git"
   },
   "keywords": [],
-  "author": "",
+  "author": "Luke Milsom - Front End Developer",
   "license": "ISC",
   "bugs": {
     "url": "https://github.com/LukeMilsom/Poker-Hand-Sorter/issues"
