+++ utils/combinationChecker.mjs
@@ -10,30 +10,16 @@ import {
   isHighCard,
 } from "./combinations/index.mjs";
 
-export const combinationChecker = (playerHand) => {
-  //definite one loop in refactor
-  const cardNumbers = playerHand.map((item) => {
-    const numberAndSuit = item.split("");
-    return numberAndSuit[0];
-  });
+import {
+  playerHandCardNumbers,
+  playerHandCardSuits,
+} from "./commonUtilities.mjs";
 
-  const cardSuits = playerHand.map((item) => {
-    const numberAndSuit = item.split("");
-    return numberAndSuit[1];
-  });
+export const combinationChecker = (playerHand) => {
+  const cardNumbers = playerHandCardNumbers(playerHand);
+  const cardSuits = playerHandCardSuits(playerHand);
 
   let rank;
-
-  //console.log(isRoyalFlush(cardNumbers, cardSuits) + "isRoyalFlush");
-  //console.log(isStraightFlush(cardNumbers, cardSuits) + "isStraightFlush");
-  //console.log(isFourOfAKind(cardNumbers, cardSuits) + "isFourOfAKind");
-  //console.log(isFullHouse(cardNumbers) + "isFullHouse");
-  //console.log(isFlush(cardSuits) + "isFLush");
-  //console.log(isStraight(cardNumbers) + "isStraight");
-  //console.log(isThreeOfAKind(cardNumbers) + " is three of a kind");
-  //console.log(isPairs(cardNumbers, 2) + "is two pairs");
-  //console.log(isPairs(cardNumbers, 1) + "is one pair");
-  //console.log(isHighCard(cardNumbers));
 
   switch (true) {
     case isRoyalFlush(cardNumbers, cardSuits):
