+++ utils/settle-tie/index.mjs
@@ -1,7 +1,6 @@
 import { settleStraightFlush } from "./settleStraightFlush.mjs";
 import { settleFourOfAKind } from "./settleFourOfAKind.mjs";
 import { settleFullHouse } from "./settleFullHouse.mjs";
-import { settleFlush } from "./settleFlush.mjs";
 import { settleStraight } from "./settleStraight.mjs";
 import { settleThreeOfAKind } from "./settleThreeOfAKind.mjs";
 import { settleTwoPairs } from "./settleTwoPairs.mjs";
@@ -13,7 +12,6 @@ export {
   settleStraightFlush,
   settleFourOfAKind,
   settleFullHouse,
-  settleFlush,
   settleTwoPairs,
   settleThreeOfAKind,
   settleHighCard,
