--- utils/settle-tie/settleFlush.mjs
@@ -1,1 +1,0 @@
-export const settleFlush = (players) => {};
