+++ utils/settle-tie/settleFourOfAKind.mjs
@@ -1,4 +1,4 @@
-import { isMatchingCardNumbers } from "../commonUtilities.mjs";
+import { usePairTieBreaker } from "../commonUtilities.mjs";
 
 export const settleFourOfAKind = (players) => {
   const matchAmount = 4;
@@ -6,27 +6,5 @@ export const settleFourOfAKind = (players) => {
   const playerOneNumbers = players.playerOne.numbers;
   const playerTwoNumbers = players.playerTwo.numbers;
 
-  let playerOneHighCard;
-  let playerTwoHighCard;
-
-  for (let i = 0; i < playerOneNumbers.length; i++) {
-    if (
-      isMatchingCardNumbers(playerOneNumbers, matchAmount, playerOneNumbers[i])
-    ) {
-      playerOneHighCard = playerOneNumbers[i];
-    }
-    if (
-      isMatchingCardNumbers(playerTwoNumbers, matchAmount, playerTwoNumbers[i])
-    ) {
-      playerTwoHighCard = playerTwoNumbers[i];
-    }
-  }
-  if (playerOneHighCard > playerTwoHighCard) {
-    return "player_one";
-  }
-  if (playerTwoHighCard > playerOneHighCard) {
-    return "player_two";
-  }
-
-  return false;
+  return usePairTieBreaker(playerOneNumbers, playerTwoNumbers, matchAmount);
 };
