+++ utils/settle-tie/settleStraightFlush.mjs
@@ -1,26 +1,8 @@
-import { sortByCardNumber } from "../commonUtilities.mjs";
+import { sortAndSettleHighCard } from "../commonUtilities.mjs";
 
-export const settleStraightFlush = (players, arrayRef) => {
-  const playerOneNumbers = sortByCardNumber(
-    players.playerOne.numbers,
-    arrayRef
-  );
-  const playerTwoNumbers = sortByCardNumber(
-    players.playerTwo.numbers,
-    arrayRef
-  );
+export const settleStraightFlush = (players) => {
+  const playerTwoNumbers = players.playerTwo.numbers;
+  const playerOneNumbers = players.playerOne.numbers;
 
-  const playerOneHighCard = playerOneNumbers.slice(-1)[0];
-  const playerTwoHighCard = playerTwoNumbers.slice(-1)[0];
-
-  const playerOneScore = arrayRef.indexOf(playerOneHighCard);
-  const playerTwoScore = arrayRef.indexOf(playerTwoHighCard);
-
-  if (playerOneScore > playerTwoScore) {
-    return "player_one";
-  }
-  if (playerTwoScore > playerOneScore) {
-    return "player_two";
-  }
-  return false;
+  return sortAndSettleHighCard(playerOneNumbers, playerTwoNumbers);
 };
